refactor to make settings page more readable
